+++ CHANGELOG.md
@@ -9,7 +9,10 @@
      * added `stubbles\webapp\auth\AuthenticationProvider`
      * added `stubbles\webapp\auth\AuthorizationProvider`
      * added `stubbles\webapp\auth\Roles`
+     * added `stubbles\webapp\auth\Token`
      * added `stubbles\webapp\auth\User`
+     * added `stubbles\webapp\auth\token\TokenAuthenticator`
+     * added `stubbles\webapp\auth\token\TokenStore`